Splitting navBar components
